Feature/next question

+++ app/service/game_service.py
@@ -15,7 +15,7 @@ from crazerace.http.instrumentation import trace
 from app.models import Game, GameMember, GameQuestion, Question
 from app.models.dto import CreateGameDTO, GameDTO, GameMemberDTO, CoordinateDTO
 from app.repository import game_repo, member_repo
-from app.service import util, question_service
+from app.service import util, question_service, game_state_util
 
 
 @trace("game_service")
@@ -26,7 +26,7 @@ def create_game(new_game: CreateGameDTO, user_id: str) -> None:
         created_at=new_game.created_at,
         members=[
             GameMember(
-                id=_new_id(), user_id=user_id, game_id=new_game.game_id, is_admin=True
+                id=util.new_id(), user_id=user_id, game_id=new_game.game_id, is_admin=True
             )
         ],
     )
@@ -35,9 +35,9 @@ def create_game(new_game: CreateGameDTO, user_id: str) -> None:
 
 @trace("game_service")
 def delete_game(game_id: str, user_id: str) -> None:
-    game = _assert_game_exists(game_id)
-    _assert_user_is_game_admin(user_id, game)
-    _assert_game_not_started(game)
+    game = game_state_util.assert_game_exists(game_id)
+    game_state_util.assert_user_is_game_admin(user_id, game)
+    game_state_util.assert_game_not_started(game)
     game_repo.delete(game)
 
 
@@ -68,21 +68,21 @@ def start_game(game_id: str, user_id: str, coordinate: CoordinateDTO) -> None:
 
 @trace("game_service")
 def add_game_member(game_id: str, user_id: str) -> None:
-    _assert_game_exists(game_id)
+    game_state_util.assert_game_exists(game_id)
     member = GameMember(id=util.new_id(), game_id=game_id, user_id=user_id)
     member_repo.add_member(member)
 
 
 @trace("game_service")
 def set_game_member_as_ready(game_id: str, member_id: str, user_id: str) -> None:
-    _assert_valid_game_member(game_id, member_id, user_id)
+    game_state_util.assert_valid_game_member(game_id, member_id, user_id)
     member_repo.set_as_ready(member_id)
 
 
 @trace("game_service")
 def leave_game(game_id: str, member_id: str, user_id: str) -> None:
-    game = _assert_game_exists(game_id)
-    member = _assert_valid_game_member(game_id, member_id, user_id)
+    game = game_state_util.assert_game_exists(game_id)
+    member = game_state_util.assert_valid_game_member(game_id, member_id, user_id)
     if game.started_at is None:
         member_repo.delete_member(member)
     else:
@@ -91,59 +91,13 @@ def leave_game(game_id: str, member_id: str, user_id: str) -> None:
 
 
 @trace("game_service")
-def _assert_game_exists(game_id: str) -> None:
-    game = game_repo.find(game_id)
-    if not game:
-        raise PreconditionRequiredError(f"Game with id={game_id} does not exit")
-    return game
-
-
-@trace("game_service")
-def _assert_valid_game_member(game_id: str, member_id: str, user_id: str) -> GameMember:
-    member = member_repo.find(member_id)
-    if not member:
-        raise PreconditionRequiredError(
-            f"Game member with id={member_id} does not exit"
-        )
-    elif member.user_id != user_id:
-        raise ForbiddenError("User ID and member ID is not related")
-    elif member.game_id != game_id:
-        raise PreconditionRequiredError(f"Member not part of game with id={game_id}")
-    return member
-
-
-@trace("game_service")
 def _find_game_and_assert_can_be_started(game_id: str, user_id: str) -> Game:
-    game = _assert_game_exists(game_id)
-    _assert_user_is_game_admin(user_id, game)
-    _assert_all_members_ready(game)
+    game = game_state_util.assert_game_exists(game_id)
+    game_state_util.assert_user_is_game_admin(user_id, game)
+    game_state_util.assert_all_members_ready(game)
     if game.started_at is not None:
         raise ConflictError(f"Game with id={game_id} is already started")
     return game
-
-
-def _assert_all_members_ready(game: Game) -> None:
-    members_ready = [member.is_ready for member in game.members]
-    if not all(members_ready):
-        raise PreconditionRequiredError(
-            f"All members are not ready in game with id={game.id}"
-        )
-
-
-def _assert_user_is_game_admin(user_id: str, game: Game) -> None:
-    for member in game.members:
-        if member.user_id == user_id and member.is_admin:
-            return
-    raise ForbiddenError(f"User is not admin on game with id={game.id}")
-
-
-def _assert_game_not_started(game: Game) -> None:
-    if game.started_at:
-        raise PreconditionRequiredError(f"Game is started")
-
-
-def _new_id() -> str:
-    return str(uuid4()).lower()
 
 
 def _map_members_to_dto(members: List[GameMember]) -> List[GameMemberDTO]:
@@ -160,9 +114,7 @@ def _map_members_to_dto(members: List[GameMember]) -> List[GameMemberDTO]:
     ]
 
 
-def _map_questions_to_game(
-    game_id: str, questions: List[Question]
-) -> List[GameQuestion]:
+def _map_questions_to_game(game_id: str, questions: List[Question]) -> List[GameQuestion]:
     return [GameQuestion(game_id=game_id, question_id=q.id) for q in questions]
